設定まわりの定義を umapita_setting.h に追い出した
